refactor: modularize game drafting flows
