+++ apps/api/src/bit_indie_api/main.py
@@ -8,7 +8,9 @@ from bit_indie_api.api.v1.routes.admin_refunds import router as admin_refunds_ro
 from bit_indie_api.api.v1.routes.admin_stats import router as admin_stats_router
 from bit_indie_api.api.v1.routes.comments import router as comments_router
 from bit_indie_api.api.v1.routes.developers import router as developers_router
-from bit_indie_api.api.v1.routes.games import router as games_router
+from bit_indie_api.api.v1.routes.game_catalog import router as game_catalog_router
+from bit_indie_api.api.v1.routes.game_drafts import router as game_drafts_router
+from bit_indie_api.api.v1.routes.game_purchases import router as game_purchases_router
 from bit_indie_api.api.v1.routes.health import router as health_router
 from bit_indie_api.api.v1.routes.purchases import router as purchases_router
 from bit_indie_api.api.v1.routes.users import router as users_router
@@ -41,7 +43,9 @@ def create_application() -> FastAPI:
     application.include_router(admin_stats_router)
     application.include_router(developers_router)
     application.include_router(users_router)
-    application.include_router(games_router)
+    application.include_router(game_catalog_router)
+    application.include_router(game_drafts_router)
+    application.include_router(game_purchases_router)
     application.include_router(comments_router)
     application.include_router(reviews_router)
     application.include_router(purchases_router)
